pyproject.toml
[tool.poetry]
name = "airflow_ecr_plugin"
version = "0.1.0"
description = "Airflow ECR plugin"
authors = ["Sandeep Aggarwal <asandeep.me@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/asandeep/airflow-ecr-plugin"
repository = "https://github.com/asandeep/airflow-ecr-plugin.git"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^2.7"
apache-airflow = {version = "^1.10", optional = true}
boto3 = "^1.10"

[tool.poetry.dev-dependencies]
pytest = [
    {version = "^4.6", python = "^2.7"},
    {version = "^5.3", python = "^3.0"}
]
moto = "^1.3"

[tool.poetry.extras]
airflow = ["apache-airflow"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
